Created search by ingredients/dishes route

+++ API/src/controllers/categoriesController.js
@@ -1,55 +1,89 @@
-const knex = require('knex');
+const AppError = require("../utils/AppError")
 
-const AppError = require('../utils/AppError');
+const knex = require("../database/knex")
 
 class CategoriesController {
 
+    /**
+     * Creates a new ingredient in the database.
+     *
+     * @param {Object} request - The request object containing the body data.
+     * @param {Object} response - The response object to send back the result.
+     * @param {string} request.body.name - The name of the ingredient to be created.
+     * @throws Will throw an error if the name is not provided or if the ingredient already exists.
+     * @returns {Object} - Returns a response object with status 201 and an empty JSON body.
+     */
     async create(request, response) {
-        const { name } = request.body;
+        const { name } = request.body
 
+        // validate name is provided and not empty
         if (!name) {
-            throw new AppError('Name is required');
+            throw new AppError("Name is required.")
         }
 
-        const category = await knex('Categories').insert({ name });
+        // check if ingredient already exists
+        const checkCategoryExists = await knex("Categories").where({ name }).first()
+        if (checkCategoryExists) {
+            throw new AppError("Category already exists.")
+        }
 
-        return response.status(201).json(category);
+        await knex("Categories").insert({ name })
+
+        return response.status(201).json()
     }
 
     async update(request, response) {
-        const { id } = request.params;
-        const { name } = request.body;
+        const { id } = request.params
+        const { name } = request.body
 
+        // validate name is provided and not empty
         if (!name) {
-            throw new AppError('Name is required');
+            throw new AppError("Category name is required.")
         }
 
-        await knex('Categories').where({ id }).update({ name });
+        // check if ingredient already exists
+        const checkCategoryExists = await knex("Categories").where({ id }).first()
+        if (!checkCategoryExists) {
+            throw new AppError("Categories not found.")
+        }
 
-        return response.json();
+        await knex("Categories").where({ id }).update({ name })
+
+        return response.status(200).json()
     }
 
     async delete(request, response) {
-        const { id } = request.params;
+        const { id } = request.params
 
-        await knex('Categories').where({ id }).delete();
+        // check if ingredient exists
+        const checkCategoryExists = await knex("Categories").where({ id }).first()
+        if (!checkCategoryExists) {
+            throw new AppError("Category not found.")
+        }
 
-        return response.json();
+        await knex("Categories").where({ id }).delete()
+
+        return response.status(200).json()
     }
 
     async index(request, response) {
-        const categories = await knex('Categories');
+        const categories = await knex("Categories").orderBy("id", "desc")
 
-        return response.json(categories);
+        return response.json(categories)
     }
 
     async show(request, response) {
-        const { id } = request.params;
+        const { id } = request.params
 
-        const category = await knex('Categories').where({ id }).first();
+        const ingredient = await knex("Categories").where({ id }).first()
 
-        return response.json(category);
+        // check if the ingredient was found
+        if (!ingredient) {
+            throw new AppError("Category not found.")
+        }
+
+        return response.json(ingredient)
     }
 }
 
-module.exports = CategoriesController;
+module.exports = CategoriesController
